Migrate palette fade mode
